Oag 107 visualise links not working as e (#65)
* Added tribe and my home IP to dev list

* Added server host to dportal url

* Support remote host in dportal url

* Fixed template requirement for dportal

* Fixed template requirement for dportal

* Fixed undefined path in cove controller

* Fixed undefined path in cove controller

* Loads of debug added

+++ src/OagBundle/Controller/CoveController.php
@@ -55,7 +55,7 @@ class CoveController extends Controller {
                     return $this->redirectToRoute('oag_default_index');
                 }
                 // else
-                if ($this->getParameter('unlink_files')) {
+                if ($this->getParameter('unlink_files') && isset($path) && file_exists($path)) {
                     unlink($path);
                 }
